pyproject.toml
[project]
name = "sbhacks12"
version = "0.1.0"
description = "VS Code extension with pygls Python language server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.25.0",
    "mss>=10.1.0",
    "numpy>=2.4.1",
    "opencv-python>=4.11.0.86",
    "twelvelabs>=1.1.0",
    "pygls>=1.3.0",
    "lsprotocol>=2023.0.0",
    "imageio[ffmpeg]>=2.37.2",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=12.1",
    "pillow>=12.1.0",
    "pywinctl>=0.4.1",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "mcp-server-demo",
]
[project.scripts]
sbhacks12-server = "server.server:main"
